Vector cached cursors ()
Use cached cursors for vector stores.

[Underwhelming benchmark
results](https://benchmarking.neo4j.today/index.html?route=_dashboards_branch-form&personal-projects=sherfert/vector-cached-cursors&personal-projects=sherfert/mydev&public-projects=dev&deselected=%5B%22statistical-fn%22+%22max%22%5D&deselected=%5B%22statistical-fn%22+%22min%22%5D&deselected=%5B%22statistical-fn%22+%22perc_25%22%5D&deselected=%5B%22statistical-fn%22+%22perc_50%22%5D&deselected=%5B%22statistical-fn%22+%22perc_75%22%5D&deselected=%5B%22statistical-fn%22+%22perc_90%22%5D&deselected=%5B%22statistical-fn%22+%22perc_95%22%5D&deselected=%5B%22statistical-fn%22+%22perc_99%22%5D&deselected=%5B%22statistical-fn%22+%22perc_99_9%22%5D&deselected=%5B%22statistical-fn%22+%22sample_size%22%5D&deselected=%5B%22threads%22+%228%22%5D&group-by=branch)

+++ community/token-api/src/main/java/org/neo4j/storageengine/api/cursor/CursorType.java
@@ -23,5 +23,5 @@ public interface CursorType {
     /**
      * @return A storage-engine unique value for this specific cursor type
      */
-    short value();
+    int value();
 }
